Slim /create and /builds/$slug to thin route shells
Both routes shrink to ~120 lines (route config + page wrapper + Suspense
shell). All view/edit body work moves into named components.

routes/create.tsx: 1483 → 118 lines. EditorShell + the in-memory
variants/cache live in `components/build-editor/editor-shell.tsx`
(1020 lines); EditorVariantBar splits out into
`components/build-editor/editor-variant-bar.tsx` (231 lines). The route
exports a `resetEditorCache()` cleanup so navigation away from /create
clears the per-build variants cache.

routes/builds.$slug.tsx: 1507 → 122 lines. New
`components/build-viewer/` folder holds the viewer body
(`build-viewer-body.tsx`), the four embed strips
(`embed-strips.tsx`), the read-only header
(`viewer-header.tsx`), social/share/actions menus, the related-builds
strip, the variant tab bar, the guide reader, and the embed shell.

Final phase of the twin-route collapse. No behavior change — every
component moved verbatim with one prop refactor (`Route.useSearch()`
calls inside EditorShell and BuildViewerBody now take `search` /
`activeIndex` as props so the components live outside the route file).
