+++ apps/web/src/components/build-viewer/build-actions-menu.tsx
@@ -1,0 +1,121 @@
+import { useNavigate } from "@tanstack/react-router"
+import { GitFork, MoreHorizontal, Trash2 } from "lucide-react"
+import { useState } from "react"
+
+import { Button } from "@/components/ui/button"
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/ui/dialog"
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu"
+import { authClient } from "@/lib/auth-client"
+import { useDeleteBuild, useForkBuild } from "@/lib/queries/build-actions"
+
+/**
+ * "More actions" overflow menu for the viewer header: Fork (anyone signed
+ * in), Delete (owner only, behind a confirm dialog).
+ */
+export function BuildActionsMenu({
+  slug,
+  name,
+  isOwner,
+}: {
+  slug: string
+  name: string
+  isOwner: boolean
+}) {
+  const { data: session } = authClient.useSession()
+  const navigate = useNavigate()
+  const fork = useForkBuild(slug)
+  const del = useDeleteBuild(slug)
+  const [confirmOpen, setConfirmOpen] = useState(false)
+
+  const onFork = () => {
+    if (!session?.user) {
+      navigate({ to: "/auth/signin" })
+      return
+    }
+    fork.mutate(undefined, {
+      onSuccess: ({ slug: newSlug }) => {
+        navigate({ to: "/builds/$slug", params: { slug: newSlug } })
+      },
+    })
+  }
+
+  const onDelete = () => {
+    del.mutate(undefined, {
+      onSuccess: () => {
+        setConfirmOpen(false)
+        navigate({ to: "/builds/mine" })
+      },
+    })
+  }
+
+  return (
+    <>
+      <DropdownMenu>
+        <DropdownMenuTrigger
+          render={
+            <Button variant="ghost" size="icon" aria-label="More actions" />
+          }
+        >
+          <MoreHorizontal />
+        </DropdownMenuTrigger>
+        <DropdownMenuContent align="end" className="min-w-40">
+          <DropdownMenuItem onClick={onFork} disabled={fork.isPending}>
+            <GitFork className="size-4" />
+            {fork.isPending ? "Forking…" : "Fork"}
+          </DropdownMenuItem>
+          {isOwner ? (
+            <>
+              <DropdownMenuSeparator />
+              <DropdownMenuItem
+                variant="destructive"
+                onClick={() => setConfirmOpen(true)}
+              >
+                <Trash2 className="size-4" />
+                Delete
+              </DropdownMenuItem>
+            </>
+          ) : null}
+        </DropdownMenuContent>
+      </DropdownMenu>
+      <Dialog open={confirmOpen} onOpenChange={setConfirmOpen}>
+        <DialogContent>
+          <DialogHeader>
+            <DialogTitle>Delete this build?</DialogTitle>
+            <DialogDescription>
+              This permanently removes “{name}”. This action cannot be undone.
+            </DialogDescription>
+          </DialogHeader>
+          <DialogFooter>
+            <Button
+              variant="outline"
+              onClick={() => setConfirmOpen(false)}
+              disabled={del.isPending}
+            >
+              Cancel
+            </Button>
+            <Button
+              variant="destructive"
+              onClick={onDelete}
+              disabled={del.isPending}
+            >
+              {del.isPending ? "Deleting…" : "Delete"}
+            </Button>
+          </DialogFooter>
+        </DialogContent>
+      </Dialog>
+    </>
+  )
+}
