+++ apps/web/src/components/build-viewer/build-not-found.tsx
@@ -1,0 +1,10 @@
+import { RouteNotFound } from "@/components/route-not-found"
+
+export function BuildNotFound() {
+  return (
+    <RouteNotFound
+      title="Build not found"
+      message="This build may have been deleted or is private."
+    />
+  )
+}
